% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boe.R
\name{get_boe}
\alias{get_boe}
\title{Fetch BoE timeseries data}
\usage{
get_boe(series, from = as.Date("1975-01-01"), to = Sys.Date())
}
\arguments{
\item{series}{a vector of BoE currency codes}

\item{from}{start date, default is 1 January 1975 (when BoE records begin)}

\item{to}{end date, default is today's date}
}
\description{
This function fetches exchange rate data from the Bank of England (BoE) database, see \url{https://www.bankofengland.co.uk/boeapps/database/}
}
\examples{
\dontrun{
get_boe(c("xumauss", "xumaers"))
get_boe("xumauss", from = as.Date("2000-01-01"), to = as.Date("2005-01-01"))
}
}
