% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boe_correlation.R
\name{get_boe_correlation}
\alias{get_boe_correlation}
\title{Calculate correlation between currency movements in BoE data}
\usage{
get_boe_correlation()
}
\arguments{
\item{date}{date value, when the forwards are from, default is today's date}

\item{three_point}{logical value, whether or not forwards should be presented as three-point estimate (minimum, most likely and maximum value)}
}
\description{
This function calculates the correlation between currency movements in BoE data
}
\examples{
\dontrun{
get_usd_forward(three_point = FALSE))
get_usd_forward(date = as.Date("2005-01-01"))
}
}
