% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boe_error_bounds.R
\name{get_boe_error_bounds}
\alias{get_boe_error_bounds}
\title{Calculate error bounds for BoE currency data}
\usage{
get_boe_error_bounds(series, steps = 12, percentile = 0.95)
}
\arguments{
\item{series}{a vector of BoE currency codes}

\item{steps}{an integer of how many steps to calculate error bounds over}

\item{percentile}{a numerical value detailing the percentile at which currency movements are evaluated}
}
\description{
This function calculates error bounds for BoE currency data. Error bounds are based off of percentage changes in a given BoE series over different intervals.
}
\examples{
\dontrun{
get_boe_error_bounds("xumauss"))
get_boe_error_bounds("xumauss", steps = 24, percentile = 0.90)
}
}
