% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usd_forward.R
\name{get_usd_forward}
\alias{get_usd_forward}
\title{Fetch BoE timeseries data}
\usage{
get_usd_forward(date = Sys.Date(), three_point = TRUE)
}
\arguments{
\item{date}{date value, when the forwards are from, default is today's date}

\item{three_point}{logical value, whether or not forwards should be presented as three-point estimate (minimum, most likely and maximum value)}
}
\description{
This function fetches exchange rate data from the Bank of England (BoE) database, see \url{https://www.bankofengland.co.uk/boeapps/database/}
}
\examples{
\dontrun{
get_usd_forward(three_point = FALSE))
get_usd_forward(date = as.Date("2005-01-01"))
}
}
