% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boe_correlation.R
\name{get_boe_correlation}
\alias{get_boe_correlation}
\title{Calculate correlation between currency movements in BoE data}
\usage{
get_boe_correlation(series, steps)
}
\arguments{
\item{series}{a vector of BoE currency codes}

\item{steps}{an integer value detailing the number of different correlation intervals to consider}
}
\description{
This function calculates the correlation between currency movements in BoE data
}
\examples{
\dontrun{
get_boe_correlation("xumauss", 12))
}
}
